Remove jQuery from it-standards.csv
jQuery is a JavaScript library that is frequently used
on sites, but we have received guidance that software
libraries are not required to get approval.

By having two versions of jQuery listed here it
implies that there are some versions of jQuery
that are not approved for use. In fact, all versions
of jQuery are approved for use since libraries do not
need approval.